requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.12.30
aiohttp>=3.8.0
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0